- move getStorageApiResourceAccessUrl to typescript-common - release it from @magda/utils
